[REPO] Split parser implementation from tests
